pressure_sensor to pressure

+++ src/embedded/Hover/pressure_check.c
@@ -6,19 +6,19 @@ Description: Loop to continuously read air pressure inside the hovercraft
 */
 
 #include <stdio.h>
-#include "pressure_sensor.h"   /* file from pressure sensor group */
+#include "pressure.h"   /* file from pressure sensor group */
 //#include <hovering_motor.h>    /* file from our own group */
 #include "pressure_check.h"
 #define PROPER_GAP 50
 #define MAX_GAP_RANGE 60
 #define MIN_GAP_RANGE 40
 
-int start_sensor_reading(){
+int start_sensor_reading() {
     handle_pressure(get_pressure());
     return 0;
 }
 
-int handle_pressure(float gap){
+int handle_pressure(int gap){
   
   if(gap > MAX_GAP_RANGE){
     // increase(10);   /* function from hovering_motor.c to increase fan speed */
